Move validate_application_error_trace_count into validators directory
