+++ tests/agent_features/test_notice_error.py
@@ -20,13 +20,19 @@ from testing_support.fixtures import (
     override_application_settings,
     reset_core_stats_engine,
     validate_application_error_event_count,
-    validate_application_error_trace_count,
     validate_transaction_error_event_count,
     validate_transaction_error_trace_count,
 )
+from testing_support.validators.validate_application_error_trace_count import (
+    validate_application_error_trace_count,
+)
+from testing_support.validators.validate_application_errors import (
+    validate_application_errors,
+)
+from testing_support.validators.validate_transaction_errors import (
+    validate_transaction_errors,
+)
 
-from testing_support.validators.validate_application_errors import validate_application_errors
-from testing_support.validators.validate_transaction_errors import validate_transaction_errors
 from newrelic.api.application import application_instance as application
 from newrelic.api.application import application_settings
 from newrelic.api.background_task import background_task
